Moved unit-tests to tests
