+++ tests/it/main.rs
@@ -1,0 +1,1 @@
+mod buffer;
